LWF Changes

+++ myDatasets.py
@@ -190,8 +190,8 @@ class ConcatDataset(torch.utils.data.Dataset):
                 for t in dataset.labels:
                     self.targets.append(t)
 
-        self.data = torch.from_numpy(np.array(self.data))
-        self.targets = torch.from_numpy(np.array(self.targets))
+        self.data = torch.from_numpy(np.array(self.data, dtype=np.float16))
+        self.targets = torch.from_numpy(np.array(self.targets, dtype=np.float16))
 
     def __getitem__(self, index):
         img, target = self.data[index], int(self.targets[index])
